pipeline-risk-factor/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.31
rapidfuzz>=3.0.0
statsmodels>=0.14.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyarrow>=14.0.0
pyyaml>=6.0
requests>=2.31.0
pytest>=7.4.0
